Change the Board Evaluation value to int
And fix a crash when moving the King

+++ ChessAndroid/app/src/main/cpp/chess/data/Board.h
@@ -23,7 +23,7 @@ public:
 	Piece *operator[](const Pos &pos);
 	const Piece *operator[](const Pos &pos) const;
 
-	float evaluate() const;
+	int evaluate() const;
 	std::unordered_map<Pos, Piece*> getAllPieces() const;
 	std::vector<Move> listAllMoves(const bool isWhite) const;
 };
